Remove unused methods moved from RedirectedURLHandler -> RedirectedURLService and update tests
